+++ omechu-app/src/app/components/restaurant/restaurant-detail/modal/ImageUploader.tsx
