+++ omechu-app/src/app/components/restaurant/restaurant-detail/modal/RatingSelector.tsx
